python/requirements.txt
jaeger-client
pyyaml